+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog for package layered_hardware_gazebo
 --------------------
 * Add posvel operation mode (position mode with profile velocity)
 * Fix velocity operation mode
+* Change parameter namespace to <robot_name>/layered_hardware_gazebo
 
 0.0.1 (2020-04-28)
 ------------------