+++ CHANGELOG.rst
@@ -2,7 +2,7 @@
 Changelog for package layered_hardware_gazebo
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-0.0.2 (next version)
+0.0.2 (2020-04-30)
 --------------------
 * Add pisition/velocity/posvel PID operation mode (effort-based position/velocity/posvel mode)
 * Add posvel operation mode (position mode with profile velocity)